requirements.txt
openai>=1.60.0
anthropic>=0.42.0
datasets>=3.0.0
python-dotenv>=1.0.0
